+++ truck-stepio/tests/output/topology.rs
@@ -1,6 +1,5 @@
 use truck_modeling::*;
 use truck_stepio::out::*;
-use truck_topology::compress::*;
 
 macro_rules! dir ( () => { concat!(env!("CARGO_MANIFEST_DIR"), "/../resources/shape/") });
 
@@ -15,8 +14,7 @@ const SOLID_JSONS: &[&str] = &[
 fn parse_solid() {
     for json_file in SOLID_JSONS.iter() {
         let json = std::fs::read(json_file).unwrap();
-        let solid: CompressedSolid<Point3, Curve, Surface> =
-            serde_json::from_reader(json.as_slice()).unwrap();
+        let solid: CompressedSolid = serde_json::from_reader(json.as_slice()).unwrap();
         let step_string =
             CompleteStepDisplay::new(StepModel::from(&solid), Default::default()).to_string();
         ruststep::parser::parse(&step_string).unwrap_or_else(|e| {
@@ -31,8 +29,7 @@ fn parse_solid() {
 fn parse_shell() {
     for json_file in SOLID_JSONS.iter() {
         let json = std::fs::read(json_file).unwrap();
-        let mut solid: CompressedSolid<Point3, Curve, Surface> =
-            serde_json::from_reader(json.as_slice()).unwrap();
+        let mut solid: CompressedSolid = serde_json::from_reader(json.as_slice()).unwrap();
         let shell = solid.boundaries.pop().unwrap();
         let step_string =
             CompleteStepDisplay::new(StepModel::from(&shell), Default::default()).to_string();
@@ -46,7 +43,7 @@ fn parse_shell() {
 
 #[test]
 fn parse_solids() {
-    let solids: Vec<CompressedSolid<Point3, Curve, Surface>> = SOLID_JSONS
+    let solids: Vec<CompressedSolid> = SOLID_JSONS
         .iter()
         .map(|json_file| {
             let json = std::fs::read(json_file).unwrap();