+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The version is of the bottom crate `truck-rendimpl`.
 
 ## Unreleased
 
+- Add the macro `truck_topology::prelude!`
+
 ### Latest `cargo upgrade`
 
 2024-10-22